Use stubber for CWE-643 tests, and test example
