+++ ql/test/query-tests/Security/CWE-312/CleartextLogging.expected
@@ -22,7 +22,7 @@ edges
 | passwords.go:127:14:127:19 | config [y] : string | passwords.go:127:14:127:21 | selection of y |
 | util.go:16:9:16:18 | selection of password : string | passwords.go:28:14:28:28 | call to getPassword |
 nodes
-| main.go:12:14:12:21 | password | semmle.label | password |
+| main.go:15:14:15:21 | password | semmle.label | password |
 | passwords.go:8:12:8:12 | definition of x : string | semmle.label | definition of x : string |
 | passwords.go:9:14:9:14 | x | semmle.label | x |
 | passwords.go:25:14:25:21 | password | semmle.label | password |
@@ -64,7 +64,7 @@ nodes
 | passwords.go:127:14:127:21 | selection of y | semmle.label | selection of y |
 | util.go:16:9:16:18 | selection of password : string | semmle.label | selection of password : string |
 #select
-| main.go:12:14:12:21 | password | main.go:12:14:12:21 | password | main.go:12:14:12:21 | password | Sensitive data returned by $@ is logged here. | main.go:12:14:12:21 | password | an access to password |
+| main.go:15:14:15:21 | password | main.go:15:14:15:21 | password | main.go:15:14:15:21 | password | Sensitive data returned by $@ is logged here. | main.go:15:14:15:21 | password | an access to password |
 | passwords.go:9:14:9:14 | x | passwords.go:30:8:30:15 | password : string | passwords.go:9:14:9:14 | x | Sensitive data returned by $@ is logged here. | passwords.go:30:8:30:15 | password | an access to password |
 | passwords.go:25:14:25:21 | password | passwords.go:25:14:25:21 | password | passwords.go:25:14:25:21 | password | Sensitive data returned by $@ is logged here. | passwords.go:25:14:25:21 | password | an access to password |
 | passwords.go:26:14:26:23 | selection of password | passwords.go:26:14:26:23 | selection of password | passwords.go:26:14:26:23 | selection of password | Sensitive data returned by $@ is logged here. | passwords.go:26:14:26:23 | selection of password | an access to password |
