+++ ql/test/query-tests/Security/CWE-643/go.mod
@@ -1,16 +1,20 @@
 module main
 
-go 1.13
+go 1.14
 
 require (
 	github.com/ChrisTrenkamp/goxpath v0.0.0-20190607011252-c5096ec8773d
 	github.com/antchfx/htmlquery v1.2.2
 	github.com/antchfx/jsonquery v1.1.2
 	github.com/antchfx/xmlquery v1.2.3
-	github.com/antchfx/xpath v1.1.4
+	github.com/antchfx/xpath v1.1.5
+	github.com/github/depstubber v0.0.0-20200413231600-392d5a70208e // indirect
 	github.com/go-xmlpath/xmlpath v0.0.0-20150820204837-860cbeca3ebc
-	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e
+	github.com/golang/groupcache v0.0.0-20200121045136-8c9f03a8e57e // indirect
 	github.com/jbowtie/gokogiri v0.0.0-20190301021639-37f655d3078f
+	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/santhosh-tekuri/xpathparser v1.0.0
-	golang.org/x/net v0.0.0-20200324143707-d3edc9973b7e
+	golang.org/x/net v0.0.0-20200226121028-0de0cce0169b // indirect
+	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
+	gopkg.in/xmlpath.v2 v2.0.0-20150820204837-860cbeca3ebc // indirect
 )
